web/requirements.txt
flask==1.0.2
Flask-WTF
redis
python-dotenv
requests